backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
solana==0.32.0
solders==0.20.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
pillow==10.1.0
requests==2.31.0